+++ src/ecs/components/level/level.ts
@@ -6,7 +6,6 @@ import {
     vec2,
     type Vec2Args,
 } from "../../../math/math";
-import { calcTransform } from "../../../math/various";
 import {
     type Comp,
     type CompList,
@@ -407,8 +406,6 @@ export function level(map: string[], opt: LevelOpt): LevelComp {
             }
 
             obj.tilePos = p;
-            // Stale, so recalculate
-            calcTransform(obj, obj.transform);
 
             if (spatialMap) {
                 insertIntoSpatialMap(obj);