+++ src/math/various.ts
@@ -1,19 +1,6 @@
 import type { GameObj } from "../types";
 import { deg2rad, Mat23, Vec2, vec2 } from "./math";
 
-export function calcTransform(obj: GameObj, tr: Mat23): Mat23 {
-    if (obj.parent) {
-        tr.setMat23(obj.parent.transform);
-    }
-    else {
-        tr.setIdentity();
-    }
-    if (obj.pos) tr.translateSelfV(obj.pos);
-    if (obj.angle) tr.rotateSelf(obj.angle);
-    if (obj.scale) tr.scaleSelfV(obj.scale);
-    return tr;
-}
-
 // convert a screen space coordinate to webgl normalized device coordinate
 export function screen2ndc(pt: Vec2, width: number, height: number, out: Vec2) {
     out.x = pt.x / width * 2 - 1;
@@ -55,4 +42,19 @@ export function getArcPts(
     }
 
     return pts;
+}
+
+export function calculateLocalTransform(obj: GameObj, tr: Mat23): Mat23 {
+    tr.setIdentity();
+    if (obj.pos) tr.translateSelfV(obj.pos);
+    if (obj.angle) tr.rotateSelf(obj.angle);
+    if (obj.scale) tr.scaleSelfV(obj.scale);
+    return tr;
+}
+
+export function calculateWorldTransform(obj: GameObj, tr: Mat23): Mat23 {
+    if (obj.parent) tr.setMat23(obj.parent.worldTransform);
+    else tr.setIdentity();
+    tr.mulSelf(obj.localTransform);
+    return tr;
 }